requirements.txt
arxiv==0.5.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
contextvars==2.4
feedparser==5.2.1
google-api-core==1.22.0
google-auth==1.19.2
google-cloud-core==1.3.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.14
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
requests==2.24.0
rfc3986==1.4.0
rsa==4.6
six==1.15.0
sniffio==1.1.0
urllib3==1.25.10
